Adds strategy parameter with the strategy name in the failure end point redirect callback.
