+++ lib/omniauth/failure_endpoint.rb
@@ -22,13 +22,18 @@ module OmniAuth
     end
 
     def raise_out!
-      raise env['omniauth.error'] || OmniAuth::Error.new(env['omniauth.error.type']) 
+      raise env['omniauth.error'] || OmniAuth::Error.new(env['omniauth.error.type'])
     end
 
     def redirect_to_failure
       message_key = env['omniauth.error.type']
-      new_path = "#{env['SCRIPT_NAME']}#{OmniAuth.config.path_prefix}/failure?message=#{message_key}#{origin_query_param}"
+      new_path = "#{env['SCRIPT_NAME']}#{OmniAuth.config.path_prefix}/failure?message=#{message_key}#{origin_query_param}#{strategy_name_query_param}"
       Rack::Response.new(["302 Moved"], 302, 'Location' => new_path).finish
+    end
+
+    def strategy_name_query_param
+      return "" unless env['omniauth.error.strategy']
+      "&strategy=#{env['omniauth.error.strategy'].name}"
     end
 
     def origin_query_param
