+++ spec/omniauth/strategy_spec.rb
@@ -1,23 +1,4 @@
-require File.expand_path('../../spec_helper', __FILE__)
-
-class ExampleStrategy
-  include OmniAuth::Strategy
-  option :name, 'test'
-  def call(env); self.call!(env) end
-  attr_reader :last_env
-  def request_phase
-    @fail = fail!(options[:failure]) if options[:failure]
-    @last_env = env
-    return @fail if @fail
-    raise "Request Phase"
-  end
-  def callback_phase
-    @fail = fail!(options[:failure]) if options[:failure]
-    @last_env = env
-    return @fail if @fail
-    raise "Callback Phase"
-  end
-end
+require 'spec_helper'
 
 def make_env(path = '/auth/test', props = {})
   {
